requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
git+https://github.com/IntelliNik/audiocraft.git@main
av==11.0.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blis==0.7.11
braceexpand==0.1.7
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.7.0
comm==0.2.0
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
datasets==2.15.0
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
demucs==4.0.1
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
dora-search==0.1.12
einops==0.7.0
encodec==0.1.1
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.9.0
flashy==0.0.2
fonttools==4.45.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
gradio==4.18.0
gradio_client==0.10.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
laion-clap==1.1.2
lameenc==1.6.3
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.2
lightning-utilities==0.10.0
llvmlite==0.41.1
loralib==0.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
num2words==0.5.13
numba==0.58.1
numpy==1.23.4
omegaconf==2.3.0
openunmix==1.2.1
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.3
peft==0.8.2
Pillow==9.3.0
platformdirs==4.0.0
pooch==1.8.0
preshed==3.0.9
proglog==0.1.10
progressbar==2.5
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-ignite==0.4.13
pytorch-lightning==2.1.2
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
resampy==0.4.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.1
ruff==0.2.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.36.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
submitit==1.5.1
sympy==1.12
terminado==0.18.0
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchlibrosa==0.1.0
torchmetrics==1.2.0
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.30.2
treetable==0.2.5
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.24.0.post1
wandb==0.16.2
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
wget==3.2
widgetsnbextension==4.0.9
xformers==0.0.22.post7
xxhash==3.4.1
yarl==1.9.3
